% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Taxes.R
\name{Taxes}
\alias{Taxes}
\title{This function consists of taxing and budgeting functions. It gives users tax data for the requested state(s),
and appends their data frames with the relative tax rates. It also allows the users to determine if a tax will
go over their budget.}
\usage{
tax_rate(state)

append_tax_rates(data, data_frame)

budget(state, price, budget)
}
\arguments{
\item{state}{a string or vector containing states in the format "AB"}

\item{data}{a vector containing states in the format "AB"}

\item{data_frame}{a data frame containing a vector of states}

\item{price}{a numeric}

\item{budget}{a numeric}
}
\description{
This function consists of taxing and budgeting functions. It gives users tax data for the requested state(s),
and appends their data frames with the relative tax rates. It also allows the users to determine if a tax will
go over their budget.
}
\examples{
 tax_rate("NC")

 append_tax_rates(mydata$states, mydata)

 budget("MA", 300, 315)


}
